+++ crates/canic-core/src/workflow/runtime/mod.rs
@@ -25,7 +25,11 @@ use crate::{
     workflow,
 };
 
-pub use nonroot::{init_nonroot_canister, post_upgrade_nonroot_canister_after_memory_init};
+pub use nonroot::{
+    init_local_nonroot_canister, init_nonroot_canister,
+    post_upgrade_local_nonroot_canister_after_memory_init,
+    post_upgrade_nonroot_canister_after_memory_init,
+};
 pub use root::{init_root_canister, post_upgrade_root_canister_after_memory_init};
 
 ///